rename folder pages to screens
