+++ src/components/InvoiceAmount.tsx
@@ -2,7 +2,7 @@ import { getDecodedLnInvoice } from '@cashu/cashu-ts'
 import Button from '@comps/Button'
 import useLoading from '@comps/hooks/Loading'
 import type { IDecodedLNInvoice } from '@model/ln'
-import { InvoiceAmountModal, InvoiceModal } from '@pages/Lightning/modal'
+import { InvoiceAmountModal, InvoiceModal } from '@screens/Lightning/modal'
 import { ThemeContext } from '@src/context/Theme'
 import { globals, highlight as hi } from '@styles'
 import { cleanUpNumericStr, vib } from '@util'
