+++ src/components/nav/Navigator.tsx
@@ -1,26 +1,26 @@
-import DetailsPage from '@comps/pages/History/Details'
-import MintInfoPage from '@comps/pages/Mints/Info'
 import type { RootStackParamList } from '@model/nav'
-import AddressbookPage from '@pages/Addressbook'
-import ContactPage from '@pages/Addressbook/Contact'
-import Dashboard from '@pages/Dashboard'
-import EncodedTokenPage from '@pages/EncodedToken'
-import HistoryPage from '@pages/History'
-import Lightning from '@pages/Lightning'
-import PayInvoicePage from '@pages/Lightning/payInvoice'
-import Mints from '@pages/Mints'
-import IntermintSwap from '@pages/Mints/IntermintSwap'
-import MintBackup from '@pages/Mints/MintBackup'
-import MintManagement from '@pages/Mints/MintManagement'
-import MintProofsPage from '@pages/Mints/Proofs'
-import QRScanPage from '@pages/QRScan.tsx'
-import SendTokenPage from '@pages/Send'
-import Settings from '@pages/Settings'
-import BackupPage from '@pages/Settings/Backup'
-import DisplaySettings from '@pages/Settings/Display'
-import SecuritySettings from '@pages/Settings/Security'
-import SuccessPage from '@pages/Success'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
+import AddressbookPage from '@screens/Addressbook'
+import ContactPage from '@screens/Addressbook/Contact'
+import Dashboard from '@screens/Dashboard'
+import EncodedTokenPage from '@screens/EncodedToken'
+import HistoryPage from '@screens/History'
+import DetailsPage from '@screens/History/Details'
+import Lightning from '@screens/Lightning'
+import PayInvoicePage from '@screens/Lightning/payInvoice'
+import Mints from '@screens/Mints'
+import MintInfoPage from '@screens/Mints/Info'
+import IntermintSwap from '@screens/Mints/IntermintSwap'
+import MintBackup from '@screens/Mints/MintBackup'
+import MintManagement from '@screens/Mints/MintManagement'
+import MintProofsPage from '@screens/Mints/Proofs'
+import QRScanPage from '@screens/QRScan.tsx'
+import SendTokenPage from '@screens/Send'
+import Settings from '@screens/Settings'
+import BackupPage from '@screens/Settings/Backup'
+import DisplaySettings from '@screens/Settings/Display'
+import SecuritySettings from '@screens/Settings/Security'
+import SuccessPage from '@screens/Success'
 import { ThemeContext } from '@src/context/Theme'
 import { useContext } from 'react'
 import { View } from 'react-native'
