+++ src/components/screens/EncodedToken.tsx
