+++ src/components/screens/Lightning/mintPanel.tsx
