+++ src/components/screens/Lightning/pageContent.tsx
@@ -7,7 +7,7 @@ import { l } from '@log'
 import { PromptModal } from '@modal/Prompt'
 import { IMintUrl, IProofSelection } from '@model'
 import { TLightningPageProps, TSendTokenPageProps } from '@model/nav'
-import { CoinSelectionModal, CoinSelectionResume } from '@pages/Lightning/modal'
+import { CoinSelectionModal, CoinSelectionResume } from '@screens/Lightning/modal'
 import { useKeyboard } from '@src/context/Keyboard'
 import { ThemeContext } from '@src/context/Theme'
 import { addToHistory } from '@store/HistoryStore'
