+++ src/components/screens/Lightning/payInvoice.tsx
@@ -10,8 +10,8 @@ import { l } from '@log'
 import { IProofSelection } from '@model'
 import { TPayLNInvoicePageProps } from '@model/nav'
 import TopNav from '@nav/TopNav'
-import AddressbookModal from '@pages/Addressbook/modal'
-import { CoinSelectionModal, CoinSelectionResume } from '@pages/Lightning/modal'
+import AddressbookModal from '@screens/Addressbook/modal'
+import { CoinSelectionModal, CoinSelectionResume } from '@screens/Lightning/modal'
 import { useKeyboard } from '@src/context/Keyboard'
 import { ThemeContext } from '@src/context/Theme'
 import { sumProofsValue } from '@src/wallet/proofs'
