+++ src/components/screens/Lightning/scannedQR.tsx
@@ -9,8 +9,8 @@ import { PromptModal } from '@modal/Prompt'
 import { IMintUrl, IProofSelection } from '@model'
 import { IDecodedLNInvoice } from '@model/ln'
 import { TQRScanPageProps } from '@model/nav'
-import { CoinSelectionModal, CoinSelectionResume } from '@pages/Lightning/modal'
 import { Picker } from '@react-native-picker/picker'
+import { CoinSelectionModal, CoinSelectionResume } from '@screens/Lightning/modal'
 import { ThemeContext } from '@src/context/Theme'
 import { addLnPaymentToHistory } from '@store/HistoryStore'
 import { getCustomMintNames, getDefaultMint, getMintName } from '@store/mintStore'
