+++ src/components/screens/Mints/Info.tsx
