+++ src/components/screens/Mints/Proofs.tsx
@@ -4,7 +4,7 @@ import { getProofsByMintUrl } from '@db'
 import { TMintProofsPageProps } from '@model/nav'
 import BottomNav from '@nav/BottomNav'
 import TopNav from '@nav/TopNav'
-import { ProofListHeader } from '@pages/Lightning/modal'
+import { ProofListHeader } from '@screens/Lightning/modal'
 import { FlashList } from '@shopify/flash-list'
 import { ThemeContext } from '@src/context/Theme'
 import { getMintCurrentKeySetId } from '@src/wallet'
