+++ src/components/screens/QRScan.tsx/Marker.tsx
