+++ src/components/screens/QRScan.tsx/index.tsx
@@ -10,7 +10,7 @@ import { l } from '@log'
 import TrustMintModal from '@modal/TrustMint'
 import { IDecodedLNInvoice } from '@model/ln'
 import { TQRScanPageProps } from '@model/nav'
-import ScannedQRDetails from '@pages/Lightning/scannedQR'
+import ScannedQRDetails from '@screens/Lightning/scannedQR'
 import { ThemeContext } from '@src/context/Theme'
 import { addToHistory } from '@store/HistoryStore'
 import { hasTrustedMint, isCashuToken, vib } from '@util'
