+++ test/components/AddressBook.test.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-return */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import AddressBook from '@comps/pages/Addressbook/Book'
+import AddressBook from '@comps/screens/Addressbook/Book'
 // import { initDb } from '@db'
 import { act, fireEvent, render, screen } from '@testing-library/react-native'
