+++ tsconfig.json
@@ -20,8 +20,8 @@
 			"@nav/*": ["src/components/nav/*"],
 			"@comps": ["src/components"],
 			"@comps/*": ["src/components/*"],
-			"@pages": ["src/components/pages"],
-			"@pages/*": ["src/components/pages/*"],
+			"@screens": ["src/components/screens"],
+			"@screens/*": ["src/components/screens/*"],
 			"@src": ["src"],
 			"@src/*": ["src/*"],
 			"@assets/*": ["assets/*"],
